suspense 모든 query 적용 및 보드 리스트에 적용
