+++ app/(routes)/boards/page.tsx
@@ -1,47 +1,20 @@
 "use client";
 
-import React from "react";
-import { useSetAtom } from "jotai";
-import { Button } from "@mui/material";
+import React, { Suspense } from "react";
 
-import Board from "_components/boards";
-import CreateBoardModal from "_components/boards/createModal";
+import BoardsList from "_components/boards/list";
 import Layout from "_components/common/layout";
-import { useGetMyBoardsQuery } from "_firebase/boards";
-import useUserAccount from "_utils/hooks/auth";
-import { modalAtom } from "_stores/modal";
+import Loading from "_components/common/loading";
 
 const BoardsPage = () => {
-  const userAccount = useUserAccount();
-
-  const { data: myBoards } = useGetMyBoardsQuery(userAccount?.uid);
-
-  const setModals = useSetAtom(modalAtom);
-
   return (
     <Layout hasHeader={true} isRequiredSignedIn={true}>
       <div className="flex justify-center items-center">
-        <div className="shadow-3xl max-w-6xl w-full p-10 h-screenWithoutHeader bg-white flex flex-col">
+        <div className="shadow-3xl max-w-6xl w-full p-10 h-screenWithoutHeader bg-white flex flex-col gap-10">
           <h2 className="text-3xl">My Boards</h2>
-          <div className="mt-10">
-            <div className="flex justify-between items-center">
-              <span>Number of My Boards: {myBoards?.length}</span>
-              <Button onClick={() => setModals((prev) => [...prev, { component: <CreateBoardModal /> }])}>
-                Create Board
-              </Button>
-            </div>
-            <div className="grid grid-cols-3 gap-10">
-              {myBoards?.map((board) => (
-                <Board
-                  boardName={board.boardName}
-                  backgroundURL={board.backgroundURL}
-                  id={board.id}
-                  ownerUID={board.ownerUID}
-                  key={board.id}
-                />
-              ))}
-            </div>
-          </div>
+          <Suspense fallback={<Loading />}>
+            <BoardsList />
+          </Suspense>
         </div>
       </div>
     </Layout>
